Add AutoMapper and Data Annotations

+++ src/ProEventos.API/Controllers/EventoController.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -29,7 +27,8 @@ namespace ProEventos.API.Controllers
       try
       {
           var eventos = await _eventoService.GetAllEventosAsync(true);
-          if(eventos == null) return NotFound("Nenhum evento encontrado.");
+          if(eventos == null) return NoContent();
+
           return Ok(eventos);
       }
       catch (Exception error)
@@ -44,7 +43,7 @@ namespace ProEventos.API.Controllers
       try
       {
           var evento = await _eventoService.GetEventoByIdAsync(id, true);
-          if(evento == null) return NotFound("Nenhum evento encontrado.");
+          if(evento == null) return NoContent();
           return Ok(evento);
       }
       catch (Exception error)
@@ -59,7 +58,7 @@ namespace ProEventos.API.Controllers
       try
       {
           var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-          if(evento == null) return NotFound("Nenhum evento por tema encontrado.");
+          if(evento == null) return NoContent();
           return Ok(evento);
       }
       catch (Exception error)
@@ -69,7 +68,7 @@ namespace ProEventos.API.Controllers
     }
 
     [HttpPost]
-    public async Task<IActionResult> Post(Evento model)
+    public async Task<IActionResult> Post(EventoDto model)
     {
       try
       {
@@ -84,7 +83,7 @@ namespace ProEventos.API.Controllers
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> Put(int id, Evento model)
+    public async Task<IActionResult> Put(int id, EventoDto model)
     {
       try
       {
@@ -103,9 +102,12 @@ namespace ProEventos.API.Controllers
     {
       try
       {
+          var evento = await _eventoService.GetEventoByIdAsync(id, true);
+          if(evento == null) return NoContent();
+
           return await _eventoService.DeleteEvento(id) 
             ? Ok("Deletado") 
-            : BadRequest("Evento não deletado.");
+            : throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento.");
       }
       catch (Exception error)
       {
